requirements.txt
ax-platform
accelerate==1.9.0
acres
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope
asttokens
attrs==25.3.0
bids-validator==1.14.7.post0
bidsschematools==1.0.10
brotlicffi
bz2file==0.98
certifi
cffi
charset-normalizer==3.4.2
ci-info
click
comm
contourpy
cycler==0.12.1
datasets==4.0.0
dcmstack
debugpy
decorator
dill==0.3.8
etelemetry
exceptiongroup
executing
filelock==3.18.0
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2
heudiconv
hf-xet==1.1.5
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==3.10
importlib_metadata
importlib_resources
ipykernel
ipython
ipython_pygments_lexers
isodate
jedi
Jinja2
joblib==1.5.1
jupyter_client
jupyter_core
kiwisolver==1.4.9
llvmlite==0.44.0
looseversion    
lxml
MarkupSafe
matplotlib==3.10.5
matplotlib-inline
mpmath
multidict==6.4.4
multiprocess==0.70.16
nest_asyncio
networkx
nibabel
nilearn==0.11.1
nipype
numba==0.61.2
numpy<2 #==2.2.6
omegaconf==2.3.0
packaging
pandas==2.3.0
parso
pathlib
pexpect
pickleshare
pillow==11.2.1
platformdirs
prompt_toolkit
propcache==0.3.2
protobuf==6.31.1
prov
psutil
ptyprocess
pure_eval
puremagic
pyarrow==20.0.0
pycparser
pydantic==2.11.7
pydantic_core==2.33.2
pydicom
pydot
Pygments
pynndescent==0.5.13
pyparsing==3.2.3
PySocks
python-dateutil
pytz==2025.2
PyYAML
pyzmq
rdflib
regex==2024.11.6
requests
safetensors==0.5.3
scikit-learn==1.7.0
scipy
sentry-sdk==2.30.0
setproctitle==1.3.6
setuptools==78.1.1
simplejson
six
smmap==5.0.2
stack_data
sympy
seaborn
threadpoolctl==3.6.0
tokenizers==0.21.2
# torch==2.7.1
# torch-geometric==2.6.1
# torch_scatter==2.1.2
# torch_sparse==0.6.18
# torchaudio==2.7.1
# torchvision==0.22.1
tornado
tqdm==4.67.1
traitlets
traits
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umap==0.1.1
umap-learn==0.5.9.post2
urllib3
wandb==0.20.1
wcwidth
wheel==0.45.1
xgboost==3.0.2
xvfbwrapper
xxhash==3.5.0
yarl==1.20.1
zipp
